refactor: rename middleware files
- Renamed `auth-middleware.ts` to `auth.ts`.
- Renamed `cors-csrf.ts` to `security.ts`.
