+++ src/auth/status.ts
@@ -1,8 +1,5 @@
 import { jsonContent } from '@/lib/helper'
-import {
-  authMiddleware,
-  authMiddlewareSchema,
-} from '@/middleware/auth-middleware'
+import { authMiddleware, authMiddlewareSchema } from '@/middleware/auth'
 import { createRoute, z } from '@hono/zod-openapi'
 import type { Context } from 'hono'
